backend/requirements.txt
pydantic-ai
pydantic
python-dotenv
matplotlib
numpy
Flask
flask-cors